requirements.txt
insightface==0.7.3
lmitf==0.1.23
matplotlib==3.10.6
networkx==3.2.1
numpy==2.3.3
openai==2.1.0
opencv_contrib_python==4.10.0.84
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
pandas==2.3.3
Pillow==11.3.0
python-dotenv==1.1.1
Requests==2.32.5
tqdm==4.67.1
wasabi==1.1.3
